+++ A2LParser/src/a2l/CompuVTab.java
@@ -12,7 +12,7 @@ import constante.ConversionType;
 
 public final class CompuVTab extends ConversionTable {
 
-    private int numberValuePairs;
+    private short numberValuePairs;
     private Map<Float, String> valuePairs;
     private String defaultValue; // DEFAULT_VALUE
 
@@ -41,7 +41,7 @@ public final class CompuVTab extends ConversionTable {
             this.name = parameters.get(2);
             this.longIdentifier = parameters.get(3);
             this.conversionType = ConversionType.getConversionType(parameters.get(4));
-            this.numberValuePairs = Integer.parseInt(parameters.get(5));
+            this.numberValuePairs = (short) Integer.parseInt(parameters.get(5));
 
             this.valuePairs = new LinkedHashMap<Float, String>();
